Add skeleton for JavaFX
