+++ src/main/java/command/FindCommand.java
@@ -1,9 +1,9 @@
-package Command;
+package command;
 
-import Task.Task;
-import Task.TaskList;
-import Dook.Ui;
-import Dook.Storage;
+import task.Task;
+import task.TaskList;
+import dook.Ui;
+import dook.Storage;
 import java.util.ArrayList;
 
 public class FindCommand extends Command {
