+++ src/main/java/command/UnmarkCommand.java
@@ -1,21 +1,20 @@
-package Command;
+package command;
 
-import Task.Task;
-import Task.TaskList;
-import Dook.Ui;
-import Dook.Storage;
-import Dook.DookException;
+import task.Task;
+import task.TaskList;
+import dook.Ui;
+import dook.Storage;
+import dook.DookException;
 
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
+    private final int positionToUnmark;
 
-    private final int positionToMark;
-
-    public MarkCommand(int positionToMark) {
-        this.positionToMark = positionToMark;
+    public UnmarkCommand(int positionToUnmark) {
+        this.positionToUnmark = positionToUnmark;
     }
 
     /**
-     * Marks a task as done.
+     * Marks a task as not done.
      *
      * @param tasks The bot TaskList.
      * @param ui The user interface.
@@ -24,24 +23,24 @@ public class MarkCommand extends Command {
      */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DookException {
-        Task toMark;
+        Task toUnmark;
         try {
-            toMark = tasks.get(positionToMark - 1);
+            toUnmark = tasks.get(positionToUnmark - 1);
         } catch (IndexOutOfBoundsException e) {
             DookException err;
             if (tasks.size() == 0) {
-                err = new DookException("Nooo! You don't have any tasks to mark :(");
+                err = new DookException("Nooo! You don't have any tasks to unmark :(");
             } else {
                 err = new DookException(String.format("Nooo! " +
                                 "You have %d tasks!" +
-                                " Valid inputs for mark is in the range [1 - %d]",
+                                " Valid inputs for unmark is in the range [1 - %d]",
                         tasks.size(), tasks.size()));
             }
             throw err;
         }
-        toMark.markAsDone();
-        ui.println("Oki! :D Good job! I've marked this task as done:");
-        ui.println(toMark.toString());
+        toUnmark.markAsNotDone();
+        ui.println("Lazy bum. >:( I've marked this task as done:");
+        ui.println(toUnmark.toString());
         storage.write(tasks);
     }
 }
