+++ src/main/java/dook/DookException.java
@@ -1,4 +1,4 @@
-package Dook;
+package dook;
 
 public class DookException extends Exception {
     public DookException(String message) {
