+++ src/main/java/dook/Launcher.java
@@ -1,0 +1,12 @@
+package dook;
+import javafx.application.Application;
+
+public class Launcher {
+
+    /**
+     * Driver function.
+     */
+    public static void main(String[] args) {
+        Application.launch(Dook.class, args);
+    }
+}
