+++ src/main/java/task/Deadline.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 import java.time.LocalDateTime;
