+++ src/main/java/task/TaskType.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 public enum TaskType {
     TODO,
